fix_4329 Added throw exception for setting K8S and docker mode
