+++ ohara-it/src/test/scala/oharastream/ohara/it/performance/TestPerformance4FtpSourceToHDFSSink.scala
@@ -21,40 +21,39 @@ import oharastream.ohara.client.configurator.v0.TopicApi.TopicInfo
 import oharastream.ohara.client.filesystem.FileSystem
 import oharastream.ohara.common.setting.ConnectorKey
 import oharastream.ohara.common.util.{CommonUtils, Releasable}
+import oharastream.ohara.connector.ftp.FtpSource
 import oharastream.ohara.connector.hdfs.sink.HDFSSink
-import oharastream.ohara.connector.jdbc.source.JDBCSourceConnector
 import oharastream.ohara.it.category.PerformanceGroup
-import org.junit.experimental.categories.Category
+import oharastream.ohara.kafka.connector.csv.CsvConnectorDefinitions
 import org.junit.Test
+import org.junit.experimental.categories.Category
 import spray.json.{JsNumber, JsString}
 
 @Category(Array(classOf[PerformanceGroup]))
-class TestPerformance4JDBCSourceToHDFSSinkOnDocker extends BasicTestPerformance4Jdbc {
-  override protected val tableName: String = s"TABLE${CommonUtils.randomString().toUpperCase()}"
+class TestPerformance4FtpSourceToHDFSSink extends BasicTestPerformance4Ftp {
+  private[this] val ftpCompletedPath = "/completed"
+  private[this] val ftpErrorPath     = "/error"
+  private[this] val (path, _, _)     = setupInputData(timeoutOfInputData)
 
   @Test
   def test(): Unit = {
+    val ftp  = ftpClient()
     val hdfs = hdfsClient()
     try {
-      createTable()
-      setupInputData(timeoutOfInputData)
-      loopInputDataThread(setupInputData)
       createTopic()
-
-      //Running JDBC Source Connector
+      loopInputDataThread(setupInputData)
+      //Running FTP Source Connector
       setupConnector(
         connectorKey = ConnectorKey.of(groupName, CommonUtils.randomString(5)),
-        className = classOf[JDBCSourceConnector].getName(),
-        settings = Map(
-          oharastream.ohara.connector.jdbc.source.DB_URL                -> JsString(url),
-          oharastream.ohara.connector.jdbc.source.DB_USERNAME           -> JsString(user),
-          oharastream.ohara.connector.jdbc.source.DB_PASSWORD           -> JsString(password),
-          oharastream.ohara.connector.jdbc.source.DB_TABLENAME          -> JsString(tableName),
-          oharastream.ohara.connector.jdbc.source.TIMESTAMP_COLUMN_NAME -> JsString(timestampColumnName),
-          oharastream.ohara.connector.jdbc.source.DB_SCHEMA_PATTERN     -> JsString(user),
-          oharastream.ohara.connector.jdbc.source.JDBC_FETCHDATA_SIZE   -> JsNumber(10000),
-          oharastream.ohara.connector.jdbc.source.JDBC_FLUSHDATA_SIZE   -> JsNumber(10000)
-        )
+        className = classOf[FtpSource].getName,
+        settings = ftpSettings
+          + (CsvConnectorDefinitions.INPUT_FOLDER_KEY -> JsString(path))
+          + (CsvConnectorDefinitions.COMPLETED_FOLDER_KEY -> JsString(
+            PerformanceTestingUtils.createFolder(ftp, ftpCompletedPath)
+          ))
+          + (CsvConnectorDefinitions.ERROR_FOLDER_KEY -> JsString(
+            PerformanceTestingUtils.createFolder(ftp, ftpErrorPath)
+          ))
       )
 
       //Running HDFS Sink Connector
@@ -70,25 +69,34 @@ class TestPerformance4JDBCSourceToHDFSSinkOnDocker extends BasicTestPerformance4
         )
       )
       sleepUntilEnd()
-    } finally Releasable.close(hdfs)
+    } finally {
+      Releasable.close(hdfs)
+      Releasable.close(ftp)
+    }
   }
 
   override protected def afterStoppingConnectors(
     connectorInfos: Seq[ConnectorInfo],
     topicInfos: Seq[TopicInfo]
   ): Unit = {
-    if (needDeleteData) {
-      //Drop table for the database
-      client.dropTable(tableName)
-
-      //Delete file from the HDFS
+    if (cleanupTestData) {
+      //Delete file for the FTP
+      val ftp  = ftpClient()
       val hdfs = hdfsClient()
       try {
+        PerformanceTestingUtils.deleteFolder(ftp, path)
+        PerformanceTestingUtils.deleteFolder(ftp, ftpCompletedPath)
+        PerformanceTestingUtils.deleteFolder(ftp, ftpErrorPath)
+
+        //Delete file from the HDFS
         topicInfos.foreach { topicInfo =>
           val path = s"${PerformanceTestingUtils.dataDir}/${topicInfo.topicNameOnKafka}"
           PerformanceTestingUtils.deleteFolder(hdfs, path)
         }
-      } finally Releasable.close(hdfs)
+      } finally {
+        Releasable.close(hdfs)
+        Releasable.close(ftp)
+      }
     }
   }
