+++ ohara-it/src/test/scala/oharastream/ohara/it/performance/TestPerformance4Oracle.scala
@@ -25,7 +25,7 @@ import org.junit.experimental.categories.Category
 import spray.json.{JsNumber, JsString}
 
 @Category(Array(classOf[PerformanceGroup]))
-class TestPerformance4OracleOnDocker extends BasicTestPerformance4Jdbc {
+class TestPerformance4Oracle extends BasicTestPerformance4Jdbc {
   override protected val tableName: String =
     s"TABLE${CommonUtils.randomString().toUpperCase()}"
