+++ ohara-it/src/test/scala/oharastream/ohara/it/performance/WithPerformanceRemoteConfigurator.scala
@@ -14,21 +14,28 @@
  * limitations under the License.
  */
 
-package oharastream.ohara.it.performance
+package oharastream.ohara.it
 
 import java.util.concurrent.TimeUnit
+
+import com.typesafe.scalalogging.Logger
 import oharastream.ohara.agent.DataCollie
 import oharastream.ohara.agent.container.ContainerClient
 import oharastream.ohara.agent.docker.DockerClient
 import oharastream.ohara.client.configurator.v0.NodeApi.{Node, State}
 import oharastream.ohara.common.util.{CommonUtils, Releasable, VersionUtils}
-import oharastream.ohara.it.{IntegrationTest, ServiceKeyHolder}
 import org.junit.{After, AssumptionViolatedException, Before}
-import org.scalatest.Matchers.contain
 import org.scalatest.Matchers._
+
 import scala.concurrent.ExecutionContext.Implicits.global
 
+/**
+  * a basic setup offering a configurator running on remote node.
+  * this stuff is also in charge of releasing the configurator after testing.
+  */
 abstract class WithPerformanceRemoteConfigurator extends IntegrationTest {
+  private[this] val log: Logger = Logger(classOf[WithRemoteConfigurator])
+
   private[this] val CONFIURATOR_NODENAME_KEY = "ohara.it.performance.configurator.node"
 
   private[this] val configuratorNodeInfo: String = sys.env.getOrElse(
@@ -61,14 +68,50 @@ abstract class WithPerformanceRemoteConfigurator extends IntegrationTest {
   private[this] val configuratorContainerName: String =
     s"${configuratorContainerKey.group()}-${configuratorContainerKey.name()}"
 
+  private[this] val k8sURL: Option[String] =
+    sys.env.get(EnvTestingUtils.K8S_MASTER_KEY).map(url => s"--k8s ${url}").orElse(Option.empty)
+
   private[this] val imageName = s"oharastream/configurator:${VersionUtils.VERSION}"
 
-  protected def nodes: Seq[Node]
-  protected def containerClient: ContainerClient
+  protected var nodes: Seq[Node]                 = _
+  protected var containerClient: ContainerClient = _
 
   @Before
   def setupConfigurator(): Unit = {
+    val k8s    = sys.env.get(EnvTestingUtils.K8S_MASTER_KEY)
+    val docker = sys.env.get(EnvTestingUtils.DOCKER_NODES_KEY)
+
+    val serviceInfo: (Seq[Node], ContainerClient) =
+      if (k8s.nonEmpty && docker.isEmpty) {
+        log.info(s"Running the K8S mode")
+        val nodes: Seq[Node]                 = EnvTestingUtils.k8sNodes()
+        val containerClient: ContainerClient = EnvTestingUtils.k8sClient()
+        (nodes, containerClient)
+      } else if (k8s.isEmpty && docker.nonEmpty) {
+        log.info(s"Running the Docker mode")
+        val nodes: Seq[Node]                 = EnvTestingUtils.dockerNodes()
+        val containerClient: ContainerClient = DockerClient(DataCollie(nodes))
+        (nodes, containerClient)
+      } else if (k8s.nonEmpty && docker.nonEmpty) {
+        throw new IllegalArgumentException("You can't setting the docker and K8S mode")
+      } else {
+        throw new IllegalArgumentException("You must setting the docker or K8S mode")
+      }
+
+    nodes = serviceInfo._1
+    containerClient = serviceInfo._2
+
     result(configuratorContainerClient.imageNames(configuratorHostname)) should contain(imageName)
+
+    val routes
+      : Map[String, String] = Map(configuratorNode.hostname -> CommonUtils.address(configuratorNode.hostname)) ++
+      nodes.map(node => node.hostname -> CommonUtils.address(node.hostname)).toMap ++
+      k8sURL
+        .map { url =>
+          val k8sMasterNodeName = url.split("http://").last.split(":").head
+          Map(k8sMasterNodeName -> CommonUtils.address(k8sMasterNodeName))
+        }
+        .getOrElse(Map.empty)
 
     result(
       configuratorContainerClient.containerCreator
@@ -76,14 +119,10 @@ abstract class WithPerformanceRemoteConfigurator extends IntegrationTest {
         .imageName(imageName)
         .portMappings(Map(configuratorPort -> configuratorPort))
         .command(
-          s"--hostname $configuratorHostname --port $configuratorPort ${otherContainerCommand()}"
+          s"--hostname $configuratorHostname --port $configuratorPort ${k8sURL.getOrElse("")}"
         )
         // add the routes manually since not all envs have deployed the DNS.
-        .routes(
-          Map(configuratorNode.hostname -> CommonUtils.address(configuratorNode.hostname)) ++
-            nodes.map(node => node.hostname -> CommonUtils.address(node.hostname)).toMap ++
-            otherRoutes()
-        )
+        .routes(routes)
         .name(configuratorContainerName)
         .create()
     )
@@ -91,10 +130,6 @@ abstract class WithPerformanceRemoteConfigurator extends IntegrationTest {
     // Wait configurator start completed
     TimeUnit.SECONDS.sleep(10)
   }
-
-  protected def otherRoutes(): Map[String, String] = Map.empty
-
-  protected def otherContainerCommand(): String = ""
 
   @After
   def releaseConfigurator(): Unit = {
